refactor(routes): colocate route-only feature files
TanStack Router의 - prefix ignore 규칙에 맞춰 홈과 서비스 상세 라우트 전용 컴포넌트, 훅을 routes 하위로 이동합니다.

현재 features 폴더에 남길 cross-route 기능이 없어 빈 features 구조를 제거하고, 공유 코드는 components/ui, components/site, lib에 남겨둡니다.

검증

- git diff --check

- pnpm typecheck

- pnpm lint

- env TMPDIR=/tmp TEMP=/tmp TMP=/tmp pnpm test (sandbox EPERM으로 escalated 재실행, 51 tests passed)

- env TMPDIR=/tmp TEMP=/tmp TMP=/tmp pnpm build (sandbox/temp socket 제약으로 escalated 재실행)
