+++ src/routes/-home/components/hero.tsx
