+++ src/routes/services/-components/detail-tabs.tsx
